XRT-1729 RISC-V build support added (#236)
* RISC-V initial prototype support

* XRT-1729 RISC-V build support update

* XRT-XRT-1729 Build flags fixup

* XRT-1729 Compiler warning fix

* XRT-1729 More cmake toolchain build updates

* RISCV build updates

* RISCV build updates

* RISCV build updates

* RISCV build updates

* SDK setup update

* XRT-1729 Build env update

* XRT-1729 Build env update

* XRT-1729 compile command fix

* XRT-1729 Build tidy

* XRT-1729 Build root fix

* XRT-1729 cmake build fix

* XRT-1729 Script fix

* XRT-1729 Build arch fix

* XRT-1729 Build system name fix

* XRT-1729 Build system name made explicit

* XRT-1729 Update to use BOARD for RISC-V build target

* XRT-1729 Update to support RISC-V OE 32 and 64 bit builds

* XRT-1729 Fix for setting 32/64 bit system

* XRT-1729 Additional riscv targets added for Pathfinder SDK

* XRT-1729 Additional riscv targets added for Pathfinder SDK #2

* XRT-1729 Code reformat

* XRT-1729 Target name updated for musl and release note added

* XRT-1729 Path fix for sdk binaries

* XRT-1729 Path fix for sdk binaries #2

* XRT-1729 Path fix for sdk binaries #3

* XRT-1729 Fix for RISC-V archir

* XRT-1729 Release note update

* XRT-1729 Update for RISC-V MUSL builds

* XRT-1729 Update for RISC-V MUSL builds

* XRT-1729 Update for RISC-V MUSL builds

* XRT-1729 Update for RISC-V MUSL builds

* XRT-1729 Update for RISC-V compil flags debug

* XRT-1729 Cmake warning removed

+++ src/c/data.c
@@ -2664,7 +2664,7 @@ static iot_data_t * iot_data_primitive_from_json (iot_json_tok_t ** tokens, cons
         uint64_t ui64;
         if (sscanf (str,"%"SCNu64, &ui64) == 1)
         {
-          ret = (ui64 <= INT64_MAX) ? iot_data_alloc_i64 (ui64) : iot_data_alloc_ui64 (ui64);
+          ret = (ui64 <= INT64_MAX) ? iot_data_alloc_i64 ((int64_t) ui64) : iot_data_alloc_ui64 (ui64);
         }
       }
       break;
